Split test_compile code from amalgamation.py
